+++ gammapy/stats/tests/test_feldman_cousins.py
@@ -80,7 +80,7 @@ def test_numerical_confidence_interval_pdfs():
     cl = 0.90
 
     x_bins = np.arange(0, n_bins_x)
-    mu_bins = np.linspace(mu_min, mu_max, mu_max / step_width_mu + 1, endpoint=True)
+    mu_bins = np.linspace(mu_min, mu_max, int(mu_max / step_width_mu + 1), endpoint=True)
 
     matrix = [stats.poisson(mu + background).pmf(x_bins) for mu in mu_bins]
 
@@ -127,7 +127,7 @@ def test_numerical_confidence_interval_values():
     cl = 0.90
 
     x_bins = np.linspace(-n_sigma * sigma, n_sigma * sigma, n_bins_x, endpoint=True)
-    mu_bins = np.linspace(mu_min, mu_max, mu_max / step_width_mu + 1, endpoint=True)
+    mu_bins = np.linspace(mu_min, mu_max, int(mu_max / step_width_mu + 1), endpoint=True)
 
     distribution_dict = dict((mu, [stats.norm.rvs(loc=mu, scale=sigma, size=5000)]) for mu in mu_bins)
